requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
firebase-admin>=6.0,<7.0
gunicorn>=21.0,<22.0
dj-database-url>=2.0,<3.0
